+++ storage/blobs.d.ts
@@ -1,0 +1,3 @@
+export const BLOBS_API_STORE_BLOB = 'storeBlob'
+export const BLOBS_API_RETRIEVE_BLOB = 'retrieveBlob'
+
