+++ storage/index.ts
@@ -1,8 +1,7 @@
 import { ipcMain, app } from 'electron'
 import { writeFileSync } from 'fs'
 import { ensureDir } from 'fs-extra'
-import { writeFile, readFile } from 'fs/promises'
-import { basename, dirname, join } from 'path'
+import { join } from 'path'
 import { handleGetLocalSpots, handleGetRemoteSpots, handleGetStoredLocalClientIds, handleRetrieveLocalClientDocs, handleRetrieveRemoteDocs, handleSaveLocalSpots, handleSaveRemoteSpots, handleStoreLocalDocs, handleStoreRemoteDocs, IDBModDoc } from './docs'
 import { getLANStoragePath } from './core'
 import { listVcrFiles, retrieveVcrs, storeVcr } from './vcrs'
@@ -10,83 +9,55 @@ import { DOCS_API_GET_LOCAL_SPOTS, DOCS_API_GET_REMOTE_SPOTS, DOCS_API_GET_STORE
 import { VIDEO_CACHE_RECORDS_API_STORE_VCR, VIDEO_CACHE_RECORDS_API_LIST_VCR_FILES, VIDEO_CACHE_RECORDS_API_RETRIEVE_VCRS } from './vcrs.d'
 import { handleRegisterClientUser } from './clients'
 import { CLIENTS_API_REGISTER_CLIENT_USER } from './clients.d'
+import { BLOBS_API_RETRIEVE_BLOB, BLOBS_API_STORE_BLOB } from './blobs.d'
+import { handleRetrieveBlob, handleStoreBlob } from './blobs'
 
 const LAN_STORAGE_PATH = getLANStoragePath(app.getPath('userData'))
 console.log('LAN_STORAGE_PATH:', LAN_STORAGE_PATH)
-const VIDEO_CACHE_PATH = join(getLANStoragePath(app.getPath('userData')), 'blobs')
+const BLOBS_PATH = join(getLANStoragePath(app.getPath('userData')), 'blobs')
 const VIDEO_CACHE_RECORDS_PATH = join(LAN_STORAGE_PATH, 'vcrs')
 const DOCS_PATH = join(LAN_STORAGE_PATH, 'docs')
 const CLIENTS_FOLDER = join(LAN_STORAGE_PATH, 'clients')
 
-const VIDEO_CACHE_API_STORE_BLOB = 'storeVideoBlob'
-const VIDEO_CACHE_API_TRY_RETRIEVE_BLOB = 'tryRetrieveVideoBlob'
-
-ipcMain.handle(VIDEO_CACHE_API_TRY_RETRIEVE_BLOB, async (_, args) => {
+ipcMain.handle(BLOBS_API_RETRIEVE_BLOB, async (_, args) => {
     if (args === 'test') {
-        return `${VIDEO_CACHE_API_TRY_RETRIEVE_BLOB} api test worked!`
+        return `${BLOBS_API_RETRIEVE_BLOB} api test worked!`
     } else if (typeof args === 'object'
         && 'blobId' in args && typeof args.blobId === 'string') {
         const { blobId } = args
-        const relativeVideoPath = dirname(blobId)
-        const fileName = basename(blobId)
-        const fullFolder = join(VIDEO_CACHE_PATH, relativeVideoPath)
-        const fullPath = join(fullFolder, fileName)
-        try {
-            const buffer = await readFile(fullPath)
-            return buffer
-        } catch (error) {
-            if (error.code === 'ENOENT') {
-                return null
-            } else {
-                // Handle other possible errors
-                console.error('An error occurred:', error.message)
-                throw error
-            }
-        }
+        return await handleRetrieveBlob(BLOBS_PATH, { blobId })
     } else {
-        throw Error(`invalid args for ${VIDEO_CACHE_API_TRY_RETRIEVE_BLOB}. Expected: { blobId: string } Got: ${JSON.stringify(args)}`)
+        throw Error(`invalid args for ${BLOBS_API_RETRIEVE_BLOB}. Expected: { blobId: string } Got: ${JSON.stringify(args)}`)
     }
 })
 
-ipcMain.handle(VIDEO_CACHE_API_STORE_BLOB, async (_, args) => {
+ipcMain.handle(BLOBS_API_STORE_BLOB, async (_, args) => {
     if (args === 'test') {
-        await ensureDir(VIDEO_CACHE_PATH)
-        const testPath = join(VIDEO_CACHE_PATH, 'mytest.txt')
+        await ensureDir(BLOBS_PATH)
+        const testPath = join(BLOBS_PATH, 'mytest.txt')
         writeFileSync(testPath, new Date(Date.now()).toISOString())
-        return `${VIDEO_CACHE_API_STORE_BLOB} api test worked! Wrote to ${testPath}`
+        return `${BLOBS_API_STORE_BLOB} api test worked! Wrote to ${testPath}`
     } else if (typeof args === 'object'
         && 'blobId' in args && typeof args.blobId === 'string'
         && 'arrayBuffer' in args && args.arrayBuffer instanceof ArrayBuffer) {
         const { blobId, arrayBuffer } = args
-        const relativeVideoPath = dirname(blobId)
-        const fileName = basename(blobId)
-        const fullFolder = join(VIDEO_CACHE_PATH, relativeVideoPath)
-        await ensureDir(fullFolder)
-        const fullPath = join(fullFolder, fileName)
-        const buffer = Buffer.from(arrayBuffer)
-        try {
-            await writeFile(fullPath, buffer)
-            return { blobId, videosFolder: VIDEO_CACHE_PATH, relativeVideoPath, fileName, fullPath, bufferLength: buffer.length }
-        } catch (error) {
-            console.error('An error occurred:', error.message)
-            throw error
-        }
+        return await handleStoreBlob(BLOBS_PATH, { blobId, arrayBuffer })
     } else {
-        throw Error(`invalid args for ${VIDEO_CACHE_API_STORE_BLOB}. Expected: {blobId: string, arrayBuffer: ArrayBuffer} Got: ${JSON.stringify(args)}`)
+        throw Error(`invalid args for ${BLOBS_API_STORE_BLOB}. Expected: {blobId: string, arrayBuffer: ArrayBuffer} Got: ${JSON.stringify(args)}`)
     }
 })
 
 ipcMain.handle(VIDEO_CACHE_RECORDS_API_STORE_VCR, async (_, args) => {
-        if (args === 'test') {
-            return `${VIDEO_CACHE_RECORDS_API_STORE_VCR} api test worked!`
-        } else if (typeof args === 'object'
-            && 'clientId' in args && typeof args.clientId === 'string'
-            && 'videoCacheRecord' in args && typeof args.videoCacheRecord === 'object') {
-            const { clientId, videoCacheRecord } = args
-            return await storeVcr(VIDEO_CACHE_RECORDS_PATH, { clientId, videoCacheRecord })
-        } else {
-            throw Error(`invalid args for ${VIDEO_CACHE_RECORDS_API_STORE_VCR}. Expected: '{ clientId: string, videoCacheRecord: { _id: string, uploadeds: boolean[] } }' Got: ${JSON.stringify(args)}`)
-        }
+    if (args === 'test') {
+        return `${VIDEO_CACHE_RECORDS_API_STORE_VCR} api test worked!`
+    } else if (typeof args === 'object'
+        && 'clientId' in args && typeof args.clientId === 'string'
+        && 'videoCacheRecord' in args && typeof args.videoCacheRecord === 'object') {
+        const { clientId, videoCacheRecord } = args
+        return await storeVcr(VIDEO_CACHE_RECORDS_PATH, { clientId, videoCacheRecord })
+    } else {
+        throw Error(`invalid args for ${VIDEO_CACHE_RECORDS_API_STORE_VCR}. Expected: '{ clientId: string, videoCacheRecord: { _id: string, uploadeds: boolean[] } }' Got: ${JSON.stringify(args)}`)
+    }
 })
 
 ipcMain.handle(VIDEO_CACHE_RECORDS_API_LIST_VCR_FILES, async (_, args) => {
